# Speed up the `Tap#formula_files` and `Tap#cask_files` methods