@apainintheneck 2024-08-26
> Wow, didn't expect `glob` to be that much faster. Maybe we need to come up with some alternative to Pathname at this point.

As you already know, doing a bunch of `Pathname` operations in a loop can be slow. We don't need to do that as often anymore because of the API which is good so I'm not sure it'd really worth it to create our own wrapper class. This is just a really bad example. It does make me wonder how some of the methods on the `Pathname` class are implemented though since it they seem to be so much slower than the equivalent `File` methods at times.