@Bo98 2024-08-26
> Edit: It looks like most of `Pathname` is implemented in C which makes me wonder even more. Could it just be that we have a lot of `Pathname` extensions that are slowing things down a lot?

It's not entirely C - most/all of the slow bits are in Ruby, particularly the ones that use regexes.