@apainintheneck 2024-08-26
You're probably right. I did see our old friend `Pathname#relative_path_from`, which is notoriously slow, is implemented in Ruby. Interestingly enough `Pathname#glob` is implemented in C but I think it might call some methods implemented in Ruby.