@Bo98 2024-08-26
Yeah `Pathname#glob` calls `Pathname#+` which is slow.

`Pathname.glob` is different and may be roughtly equivelent to what we have now.